requirements.txt
fastapi==0.115.12
uvicorn==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.51.2
httpx==0.27.2
feedparser==6.0.11
requests==2.32.3
rich==13.9.4
typer==0.16.0
python-dateutil==2.9.0.post0
